Split the root di module directory in each feature di directory. Moving each module under each feature directory. More orginized this way
