+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/marketplace/misc/simplescreen1/ui/SimpleScreen1View.kt
@@ -1,4 +1,4 @@
-package com.macaosoftware.component.navigationcompose.demo.marketplace.misc.simplescreen
+package com.macaosoftware.component.navigationcompose.demo.marketplace.misc.simplescreen1.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -15,8 +15,8 @@ import androidx.compose.ui.unit.dp
 import com.macaosoftware.component.util.BackPressHandler
 
 @Composable
-fun SimpleScreenView(
-    viewModel: SimpleScreenViewModel,
+fun SimpleScreen1View(
+    viewModel: SimpleScreen1ViewModel,
     modifier: Modifier = Modifier.fillMaxSize()
 ) {
     BackPressHandler {
@@ -28,6 +28,7 @@ fun SimpleScreenView(
             .verticalScroll(rememberScrollState()),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
+        Text("This is another screen render type")
         Button(
             modifier = Modifier.padding(vertical = 40.dp),
             onClick = {
