+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/serverui/di/ServerUiModule.kt
@@ -1,10 +1,12 @@
-package com.macaosoftware.component.navigationcompose.demo.di
+package com.macaosoftware.component.navigationcompose.demo.serverui.di
 
 import com.macaosoftware.component.core.DestinationRendersRegistry
 import com.macaosoftware.component.core.MacaoDestinationRendersRegistry
 import com.macaosoftware.component.navigationcompose.demo.marketplace.notfound.DestinationRenderNotFound
 import com.macaosoftware.component.navigationcompose.demo.marketplace.notfound.RootDestinationRenderNotFound
+import com.macaosoftware.component.navigationcompose.demo.serverui.data.ServerUiRemoteService
 import com.macaosoftware.component.navigationcompose.demo.serverui.domain.usecase.LoadChildrenDestinationUseCase
+import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 internal val serverUiModule = module {
@@ -17,4 +19,6 @@ internal val serverUiModule = module {
             rootDestinationRenderNotFound = RootDestinationRenderNotFound(),
         )
     }
+
+    singleOf(::ServerUiRemoteService)
 }
