+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/serverui/http/HttpClientBuilder.kt
@@ -1,4 +1,4 @@
-package com.macaosoftware.component.navigationcompose.demo.di.http
+package com.macaosoftware.component.navigationcompose.demo.serverui.http
 
 import io.ktor.client.*
 import io.ktor.client.plugins.*
